Tweak code to pass SCIM CRUD tests
