+++ packages/back-end/src/scim/users/getUser.ts
@@ -18,11 +18,30 @@ export async function getUser(req: ScimGetRequest, res: Response) {
 
   const orgUser = org.members.find((member) => member.id === user?.id);
 
-  // TODO: I think we need to return the user object in this case, but with active set to false
+  // TODO: Create a function to map Growthbook users to SCIM users that we can use here and in listUsers
   if (!orgUser) {
-    return res.status(404).json({
-      schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-      detail: "User is not a part of the organization",
+    return res.status(200).json({
+      schemas: ["urn:ietf:params:scim:schemas:core:2.0:User"],
+      id: req.params.id,
+      userName: user.email,
+      name: {
+        displayName: user.name,
+        givenName: user.name?.split(" ")[0],
+        familyName: user.name?.split(" ")[1],
+      },
+      active: false,
+      emails: [
+        {
+          primary: true,
+          value: user.email,
+          type: "work",
+          display: user.email,
+        },
+      ],
+      groups: [],
+      meta: {
+        resourceType: "User",
+      },
     });
   }
